+++ core/discotron.js
@@ -1,9 +1,9 @@
-const BotSettings = require("./classes/bot-settings.js");
-const Repository = require("./classes/repository.js");
-const Guild = require("./classes/guild.js");
-const Plugin = require("./classes/plugin.js");
-const Owner = require("./classes/owner.js");
-const SpamUser = require("./classes/spam-user.js");
+const BotSettings = require("./models/bot-settings.js");
+const Repository = require("./models/repository.js");
+const Guild = require("./models/guild.js");
+const Plugin = require("./models/plugin.js");
+const Owner = require("./models/owner.js");
+const SpamUser = require("./models/spam-user.js");
 const Logger = require("./utils/logger.js");
 const Login = require("./login.js");
 const db = require("./database/crud.js");
