+++ core/models/bot-settings.js
