+++ core/models/owner.js
