+++ core/models/plugin.js
